No longer setting minimal memory per thread
I had previously assumed that minimizing the amount of memory per thread would be valuable, since our system runs so lean, I didn't need megabytes per thread.  Unfortunately this caused the tests on Linux to StackOverflow when going through relatively ordinary numbers of stackframes.  I was so perplexed.  But when I used "Git bisect", I was able to narrow it down to the commit where I added -Xss 180k (set the threads in the JVM to only use 180k of memory each).