Fine tuning, tests, datalist for projects
It turns out that HTML5 has a better searchable option - "datalist".  It provides a list of values you can search through.  Select, for all its charms, only lets you search for entries if the text matches from the beginning - but datalist will match the entered text anywhere within, no javascript needed!

That said, to make it a bit nicer, added a touch of JavaScript to require that entered values match one of the items in list.  Without that, the system would let the user enter an invalid value for the project name, which the system would immediately kick back as a 500 error - not nearly as pretty.

Unrelated, looked around through the codebase for any areas which didn't have sufficient test coverage and added a few.  Ironed out a few mild kinks.