Fixing bugs, improved tests, style
Thanks to Steve Hunter for a whirlwind exploratory test session which found 2 real bugs in the span of about 10 minutes.  Way to go!

Specifically:
1. 24 hours was not accepted as a valid time for an entry (fixed by changing "less-than" to "less-than-or-equal" on some code)
2. When a time entry had 23 hours and then we edit it to be 22 hours, system showed complaint about too many hours for time entries that day. (fixed by properly subtracting the old time entry's time before calculating sum of hours for the day, when editing)

Added a new item to the general header on each page - a timestamp, so we can tell whether we're looking at the same unchanged page or not.  Updated the tests in TimeEntryUIValidationTests to suit.