Modified MessageAPI to handle custom messages
Now it is possible to send MessageAPI any custom message you need to show.  This is better than immediately returning from a post with a terrible 500 page that won't allow a back button to work.  Making this change did cause wide-ranging effects, obviously.  For one thing, there's no longer any exceptions making it past the auth layer of the API except for those in InexactInputExceptions

style and analysis changes