Significant ViewTime refactors, tdd on calc hours for week
Pulling pretty much everything in the ViewTimeAPI into the companion object, since the only value needed is the ServerData.  This will make testing easier if I need to focus on a particular method, since everything required will just be what goes in as parameters.

New feature to show calculatured hours per week, which will be used to show the user how many hours are currently entered for Sunday through Saturday of the week they are in.

Attaching weekly hours to each time entry day

css styling