Flattening layers, part 1
This set of commits is to see whether we can flatten the Time Recording Utility and its persistence layer beneath.  We're not getting much value from the design, and in fact we're getting anti-value, as I am finding situations where a utility needs to be able to access different parts of the database than the current abstraction easily allows.  For example, when I delete an employee, I need access to the authentication utilities (to see whether a user was created for the employee) and the time recording utility (to access details about the employee).  To date, I've been creating utilities that sit above the ordinary utilities to access these merged needs, but it occurred to me that a much simpler approach would be to remove the whole concept of splitting up the utility from its persistence, and letting the utility have direct access to any parts of the database it needs.

This first step, I'm just changing the signature of the time recording utilities to receive the database class, so we can create data access instances off it any way we want.

All tests still passing